+++ lib/src/models/santali_calendar_day.dart
@@ -2,12 +2,14 @@ class SantaliCalendarDay {
   final int day;
   final DateTime date;
   final bool isToday;
+  final String weekDay;
   final bool isCurrentMonth;
 
   const SantaliCalendarDay({
     required this.day,
     required this.date,
     required this.isToday,
+    required this.weekDay,
     required this.isCurrentMonth,
   });
 }