+++ test/santali_calendar_test.dart
@@ -1,5 +1,4 @@
 import 'package:santali_calendar/santali_calendar.dart';
-import 'package:santali_calendar/src/constants/weeks.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -62,13 +61,11 @@ void main() {
       print('\n-------- Current Month ----------\n');
 
       // Get month by date
-      final month = calendar.getMonthByDate(DateTime.utc(2027, 8, 2));
+      final month = calendar.getMonthByDate(DateTime.now());
       print('Month Start: ${month.startDate}');
       print('Month End: ${month.endDate}');
-      for (final weekDay in SantaliWeekDay.values) {
-        for (final day in month.calendar[weekDay]!) {
-          print('${weekDay.name}: ${day?.date}');
-        }
+      for (final day in month.days) {
+        print('Day: ${day?.day} - Weekday: ${day?.weekDay}');
       }
     });
   });